exam-factory/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-docx==1.1.2
pdfplumber==0.11.4
httpx==0.27.0
aiosmtplib==3.0.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
jinja2==3.1.4
